refactor: move map optional token
