+++ frontend/src/lib/utils/sns.utils.ts
@@ -12,10 +12,7 @@ import type {
   QuerySnsMetadata,
   QuerySnsSwapState,
 } from "$lib/types/sns.query";
-import {
-  IcrcMetadataResponseEntries,
-  type IcrcTokenMetadataResponse,
-} from "@dfinity/ledger";
+import { mapOptionalToken } from "$lib/utils/icrc-tokens.utils";
 import { AccountIdentifier, SubAccount } from "@dfinity/nns";
 import { Principal } from "@dfinity/principal";
 import type {
@@ -25,7 +22,7 @@ import type {
   SnsSwapDerivedState,
 } from "@dfinity/sns";
 import { fromNullable } from "@dfinity/utils";
-import { isNullish, isPngAsset } from "./utils";
+import { isPngAsset } from "./utils";
 
 type OptionalSnsSummarySwap = Omit<SnsSummarySwap, "params"> & {
   params?: SnsParams;
@@ -96,41 +93,6 @@ const mapOptionalMetadata = ({
     name: nullishName,
     description: nullishDescription,
   } as SnsSummaryMetadata;
-};
-
-/**
- * Token metadata is given only if the properties NNS-dapp needs (name, symbol and fee) are defined.
- */
-export const mapOptionalToken = (
-  response: IcrcTokenMetadataResponse
-): IcrcTokenMetadata | undefined => {
-  const nullishToken: Partial<IcrcTokenMetadata> = response.reduce(
-    (acc, [key, value]) => {
-      switch (key) {
-        case IcrcMetadataResponseEntries.SYMBOL:
-          acc = { ...acc, ...("Text" in value && { symbol: value.Text }) };
-          break;
-        case IcrcMetadataResponseEntries.NAME:
-          acc = { ...acc, ...("Text" in value && { name: value.Text }) };
-          break;
-        case IcrcMetadataResponseEntries.FEE:
-          acc = { ...acc, ...("Nat" in value && { fee: value.Nat }) };
-      }
-
-      return acc;
-    },
-    {}
-  );
-
-  if (
-    isNullish(nullishToken.name) ||
-    isNullish(nullishToken.symbol) ||
-    isNullish(nullishToken.fee)
-  ) {
-    return undefined;
-  }
-
-  return nullishToken as IcrcTokenMetadata;
 };
 
 /**
